chore: try to fix lifecycle scripts
`lerna publish` has been acting up, and I think it's because the lifecycle scripts don't do enough cleanup, or we're running them in the wrong order, or something.  so here's this.

it causes an "extra" rebuild between version and publish, but I think it's better to be safe than sorry here.