fix(docutils): upgrades for typedoc@0.24.x
TypeDoc v0.24.0 made some breaking changes which we were bit by.  Of note:

1. Plugins are no longer automatically discovered. ☹️
2. The `logger` option is no longer recognized.
3. `typedoc-plugin-resolve-crossmodule-references` is no longer needed.

I've modified `@appium/docutils` to smooth this over.

In addition, I've changed the behavior of `appium-docs init` to create a `typedoc.json` which _will_ consider `README.md`, as this is the most likely scenario for extensions.  I've changed the monorepo's config to explicitly _not_ use `README.md`.

_Note: This changeset does not contain necessary `package-lock.json` changes and should fail if run in isolation.